+++ core/src/elements/ons-page.js
@@ -16,8 +16,8 @@ limitations under the License.
 */
 
 import util from 'ons/util';
-import ons from 'ons/ons';
 import internal from 'ons/internal';
+import autoStyle from 'ons/autostyle';
 import ModifierUtil from 'ons/internal/modifier-util';
 import BaseElement from 'ons/base-element';
 import DeviceBackButtonDispatcher from 'ons/device-back-button-dispatcher';
@@ -59,6 +59,12 @@ const nullToolbarElement = document.createElement('ons-toolbar');
  *
  *   ...
  * </ons-page>
+ *
+ *
+ * // Infinite Scroll handler
+ * page.onInfiniteScroll = function(done) {
+ *   // load more content and call done
+ * };
  */
 class PageElement extends BaseElement {
 
@@ -114,6 +120,21 @@ class PageElement extends BaseElement {
    *   [ja]スタイル定義をカスタマイズするための名前を指定します。[/ja]
    */
 
+  /**
+   * @attribute on[-]infinite[-]scroll
+   * @type {String}
+   * @description
+   *   [en]Path of the function to be executed on infinite scrolling. Example: app.loadData[/en]
+   *   [ja]機能スクロール上で実行されている関数のパス。例：app.loadData[/ja]
+   */
+
+  /**
+   * @property onInfiniteScroll
+   * @description
+   *  [en]Function to be executed on infinite scroll. [/en]
+   *  [ja]機能スクロール上で実行されている関数。[/ja]
+   */
+
   createdCallback() {
     this.classList.add('page');
 
@@ -144,6 +165,9 @@ class PageElement extends BaseElement {
     }
 
     this._tryToFillStatusBar();
+
+    const infiniteScroll = this.getAttribute('on-infinite-scroll') || this.getAttribute('oninfinitescroll');
+    this.attributeChangedCallback('oninfinitescroll', null, infiniteScroll);
   }
 
   /**
@@ -158,6 +182,33 @@ class PageElement extends BaseElement {
    */
   set isShown(value) {
     this._isShown = value;
+  }
+
+  set onInfiniteScroll(value) {
+    if (value === null) {
+      this._onInfiniteScroll = null;
+      this._contentElement.removeEventListener('scroll', this._boundOnScroll);
+      return;
+    }
+    if (!(value instanceof Function)) {
+      throw new Error('onInfiniteScroll must be a function or null');
+    }
+    if (!this._onInfiniteScroll) {
+      this._infiniteScrollLimit = 0.9;
+      this._boundOnScroll = this._onScroll.bind(this);
+      this._contentElement.addEventListener('scroll', this._boundOnScroll);
+    }
+    this._onInfiniteScroll = value;
+  }
+
+  _onScroll() {
+    const c = this._contentElement,
+      overLimit = (c.scrollTop + c.clientHeight) / c.scrollHeight >= this._infiniteScrollLimit;
+
+    if (this._onInfiniteScroll && !this._loadingContent && overLimit) {
+      this._loadingContent = true;
+      this._onInfiniteScroll(() => this._loadingContent = false);
+    }
   }
 
   /**
@@ -199,27 +250,13 @@ class PageElement extends BaseElement {
   /**
    * @return {Boolean}
    */
-  _hasToolbarElement() {
-    return !!util.findChild(this, 'ons-toolbar');
-  }
-
-  /**
-   * @return {Boolean}
-   */
   _canAnimateToolbar() {
-    const toolbar = util.findChild(this, 'ons-toolbar');
-    if (toolbar) {
+    if (util.findChild(this, 'ons-toolbar')) {
       return true;
     }
-
-    const elements = this._contentElement.children;
-    for (let i = 0; i < elements.length; i++) {
-      if (elements[i].nodeName.toLowerCase() === 'ons-toolbar' && !elements[i].hasAttribute('inline')) {
-        return true;
-      }
-    }
-
-    return false;
+    return !!util.findChild(this._contentElement, (e) => {
+      return e.nodeName.toLowerCase() === 'ons-toolbar' && !e.hasAttribute('inline');
+    });
   }
 
   /**
@@ -269,15 +306,8 @@ class PageElement extends BaseElement {
    * @param {HTMLElement} element
    */
   _registerBottomToolbar(element) {
-    if (!util.findChild(this, '.page__status-bar-fill')) {
-      const fill = document.createElement('div');
-      fill.classList.add('page__bottom-bar-fill');
-      fill.style.width = '0px';
-      fill.style.height = '0px';
-
-      this.insertBefore(fill, this.children[0]);
-      this.insertBefore(element, null);
-    }
+    this.classList.add('page-with-bottom-toolbar');
+    this.appendChild(element);
   }
 
   attributeChangedCallback(name, last, current) {
@@ -287,27 +317,31 @@ class PageElement extends BaseElement {
       this._isMuted = this.hasAttribute('_muted');
     } else if (name === '_skipinit') {
       this._skipInit = this.hasAttribute('_skipinit');
+    } else if (name.match(/on-?infinite-?scroll/i)) {
+      if (current === null) {
+        this.onInfiniteScroll = null;
+      } else {
+        this.onInfiniteScroll = (done) => {
+          const f = util.findFromPath(current);
+          this.onInfiniteScroll = f;
+          f(done);
+        };
+      }
     }
   }
 
   _compile() {
-    ons._autoStyle.prepare(this);
+    autoStyle.prepare(this);
 
-    const background = document.createElement('div');
-    background.classList.add('page__background');
+    const background = util.create('.page__background');
+    const content = util.create('.page__content');
 
-    const content = document.createElement('div');
-    content.classList.add('page__content');
-
-    while (this.childNodes[0]) {
-      content.appendChild(this.childNodes[0]);
+    while (this.firstChild) {
+      content.appendChild(this.firstChild);
     }
 
-    const fragment = document.createDocumentFragment();
-    fragment.appendChild(background);
-    fragment.appendChild(content);
-
-    this.appendChild(fragment);
+    this.appendChild(background);
+    this.appendChild(content);
 
     ModifierUtil.initModifier(this, scheme);
 
@@ -317,13 +351,11 @@ class PageElement extends BaseElement {
   _registerExtraElement(element) {
     let extra = util.findChild(this, '.page__extra');
     if (!extra) {
-      extra = document.createElement('div');
-      extra.classList.add('page__extra');
-      extra.style.zIndex = '10001';
-      this.insertBefore(extra, null);
+      extra = util.create('.page__extra', {zIndex: 10001});
+      this.appendChild(extra);
     }
 
-    extra.insertBefore(element, null);
+    extra.appendChild(element);
   }
 
   _tryToFillStatusBar() {
@@ -331,28 +363,9 @@ class PageElement extends BaseElement {
       .then(() => {
         let fill = this.querySelector('.page__status-bar-fill');
 
-        if (fill instanceof HTMLElement) {
-          return fill;
-        }
+        if (!fill) {
+          fill = util.create('.page__status-bar-fill');
 
-        fill = document.createElement('div');
-        fill.classList.add('page__status-bar-fill');
-        fill.style.width = '0px';
-        fill.style.height = '0px';
-
-        let bottomBarFill;
-
-        for (let i = 0; i < this.children.length; i++) {
-          if (this.children[i].classList.contains('page__bottom-bar-fill')) {
-            bottomBarFill = this.children[i];
-            break;
-          }
-        }
-
-        if (bottomBarFill) {
-          this.insertBefore(fill, bottomBarFill.nextSibling);
-        }
-        else {
           this.insertBefore(fill, this.children[0]);
         }
 
@@ -360,9 +373,7 @@ class PageElement extends BaseElement {
       })
       .catch(() => {
         const el = this.querySelector('.page__status-bar-fill');
-        if (el instanceof HTMLElement) {
-          el.remove();
-        }
+        el && el.remove();
       });
   }