+++ core/src/elements/ons-tabbar/index.js
@@ -17,6 +17,7 @@ limitations under the License.
 
 import util from 'ons/util';
 import internal from 'ons/internal';
+import autoStyle from 'ons/autostyle';
 import ModifierUtil from 'ons/internal/modifier-util';
 import AnimatorFactory from 'ons/internal/animator-factory';
 import BaseElement from 'ons/base-element';
@@ -219,7 +220,7 @@ class TabbarElement extends BaseElement {
   }
 
   _compile() {
-    ons._autoStyle.prepare(this);
+    autoStyle.prepare(this);
 
     if (this.getAttribute('position') === 'auto') {
       this.setAttribute('position', ons.platform.isAndroid() ? 'top' : 'bottom');
@@ -395,7 +396,6 @@ class TabbarElement extends BaseElement {
    * @param {String} [options.animation]
    * @param {Function} [options.callback]
    * @param {Object} [options.animationOptions]
-   * @param {Boolean} options._removeElement
    * @param {Number} options.selectedTabIndex
    * @param {Number} options.previousTabIndex
    * @return {Promise} Resolves to the new page element.
@@ -413,13 +413,7 @@ class TabbarElement extends BaseElement {
 
       animator.apply(element, oldPageElement, options.selectedTabIndex, options.previousTabIndex, () => {
         if (oldPageElement !== internal.nullElement) {
-          if (options._removeElement) {
-            rewritables.unlink(this, oldPageElement, pageElement => {
-              pageElement._destroy();
-            });
-          } else {
-            oldPageElement.style.display = 'none';
-          }
+          oldPageElement.style.display = 'none';
         }
 
         element.style.display = 'block';
@@ -545,8 +539,7 @@ class TabbarElement extends BaseElement {
           }
         },
         previousTabIndex: previousTabIndex,
-        selectedTabIndex: selectedTabIndex,
-        _removeElement: removeElement
+        selectedTabIndex: selectedTabIndex
       };
 
       if (options.animation) {